debug handle for kind deploy

+++ scripts/deploy_on_kind.sh
@@ -39,7 +39,10 @@ kubectl apply -k manifests/kustomize/overlays/db
 kubectl patch deployment -n "$MR_NAMESPACE" model-registry-deployment \
 --patch '{"spec": {"template": {"spec": {"containers": [{"name": "rest-container", "image": "'$IMG'", "imagePullPolicy": "IfNotPresent"}]}}}}'
 
-kubectl wait --for=condition=available -n "$MR_NAMESPACE" deployment/model-registry-db --timeout=5m
+if ! kubectl wait --for=condition=available -n "$MR_NAMESPACE" deployment/model-registry-db --timeout=5m ; then
+    kubectl events -A
+    exit 1
+fi
 
 kubectl delete pod -n "$MR_NAMESPACE" --selector='component=model-registry-server'
